+++ resources/migrations/20190827055046-no-license-expiration.up.sql
@@ -1,0 +1,11 @@
+ALTER TABLE license
+  DROP COLUMN start,
+  DROP COLUMN endt;
+--;;
+ALTER TABLE resource_licenses
+  DROP COLUMN start,
+  DROP COLUMN endt;
+--;;
+ALTER TABLE workflow_licenses
+  DROP COLUMN start,
+  DROP COLUMN endt;