+++ resources/sql/queries.sql
@@ -249,12 +249,10 @@ WHERE appid = :application-id;
 
 -- :name create-license! :insert
 INSERT INTO license
-(ownerUserId, modifierUserId, title, type, textcontent, attachmentId, endt)
+(ownerUserId, modifierUserId, title, type, textcontent, attachmentId)
 VALUES
 (:owneruserid, :modifieruserid, :title, :type::license_type, :textcontent,
-/*~ (if (:attachmentId params) */ :attachmentId /*~*/ NULL /*~ ) ~*/,
-/*~ (if (:end params) */ :end /*~*/ NULL /*~ ) ~*/
-);
+/*~ (if (:attachmentId params) */ :attachmentId /*~*/ NULL /*~ ) ~*/);
 
 -- :name set-license-state! :!
 UPDATE license
@@ -317,22 +315,8 @@ INSERT INTO resource_licenses
 VALUES
 (:resid, :licid);
 
--- TODO: only used in test data; consider removing
--- :name set-resource-license-validity! :insert
--- :doc set license expiration
-UPDATE resource_licenses rl
-SET start = :start, endt = :end
-WHERE rl.licid = :licid;
-
--- TODO: only used in test data; consider removing
--- :name set-workflow-license-validity! :insert
--- :doc set license expiration
-UPDATE workflow_licenses wl
-SET start = :start, endt = :end
-WHERE wl.licid = :licid;
-
 -- :name get-workflow-licenses :? :*
-SELECT licid, start, endt as "end"
+SELECT licid
 FROM workflow_licenses
 WHERE wfid = :wfid
 
@@ -364,12 +348,12 @@ FROM workflow wf;
 -- - Gets application licenses by workflow and catalogue item ids
 -- - :wfid workflow id for workflow licenses
 -- - :items vector of catalogue item ids for resource licenses
-SELECT lic.id, lic.title, lic.type, lic.textcontent, wl.start, wl.endt as "end", lic.enabled, lic.archived
+SELECT lic.id, lic.title, lic.type, lic.textcontent, lic.enabled, lic.archived
 FROM license lic
 INNER JOIN workflow_licenses wl ON lic.id = wl.licid
 WHERE wl.wfid = :wfid
 UNION
-SELECT lic.id, lic.title, lic.type, lic.textcontent, rl.start, rl.endt as "end", lic.enabled, lic.archived
+SELECT lic.id, lic.title, lic.type, lic.textcontent, lic.enabled, lic.archived
 FROM license lic
 INNER JOIN resource_licenses rl ON lic.id = rl.licid
 INNER JOIN catalogue_item item ON (item.resid = rl.resid)
@@ -377,17 +361,17 @@ WHERE item.id IN (:v*:items)
 ORDER BY id;
 
 -- :name get-resource-licenses :? :*
-SELECT lic.id, lic.title, lic.type, lic.textcontent, rl.start, rl.endt as "end", lic.enabled, lic.archived
+SELECT lic.id, lic.title, lic.type, lic.textcontent, lic.enabled, lic.archived
 FROM license lic
 INNER JOIN resource_licenses rl ON lic.id = rl.licid
 WHERE rl.resid = :id;
 
 -- :name get-all-licenses :? :*
-SELECT lic.id, lic.title, lic.type, lic.textcontent, lic.start, lic.endt as "end", lic.enabled, lic.archived, lic.attachmentid
+SELECT lic.id, lic.title, lic.type, lic.textcontent, lic.enabled, lic.archived, lic.attachmentid
 FROM license lic;
 
 -- :name get-license :? :1
-SELECT lic.id, lic.title, lic.type, lic.textcontent, lic.start, lic.endt as "end", lic.enabled, lic.archived, lic.attachmentid
+SELECT lic.id, lic.title, lic.type, lic.textcontent, lic.enabled, lic.archived, lic.attachmentid
 FROM license lic
 WHERE lic.id = :id;