+++ src/clj/rems/api/licenses.clj
@@ -35,11 +35,9 @@
       :summary "Get licenses"
       :roles #{:handler :owner}
       :query-params [{disabled :- (describe s/Bool "whether to include disabled licenses") false}
-                     {expired :- (describe s/Bool "whether to include expired licenses") false}
                      {archived :- (describe s/Bool "whether to include archived licenses") false}]
       :return Licenses
       (ok (licenses/get-all-licenses (merge (when-not disabled {:enabled true})
-                                            (when-not expired {:expired false})
                                             (when-not archived {:archived false})))))
 
     (GET "/:license-id" []